requirements.txt
google-play-scraper
pandas
python-dotenv
streamlit
app-store-scraper
